Refactor checking out node modules into it's own function / module
